Extract module for template rendering
